move helpers + voronoi as component
