+++ web/locales/en/plugin__netobserv-plugin.json
@@ -149,9 +149,6 @@
   "Switch between one way / back and forth filtering": "Switch between one way / back and forth filtering",
   "One way shows traffic strictly as defined per your filters": "One way shows traffic strictly as defined per your filters",
   "Back and forth shows traffic according to your filters, plus the related return traffic": "Back and forth shows traffic according to your filters, plus the related return traffic",
-  "Source": "Source",
-  "Destination": "Destination",
-  "Common": "Common",
   "Filter already exists": "Filter already exists",
   "Hide filters": "Hide filters",
   "Show {{count}} filters": "Show {{count}} filters",
@@ -160,6 +157,8 @@
   "Collapse": "Collapse",
   "Expand": "Expand",
   "Quick filters": "Quick filters",
+  "Source": "Source",
+  "Destination": "Destination",
   "Step {{index}}/{{count}}": "Step {{index}}/{{count}}",
   "Step {{index}}/{{count}}_plural": "Step {{index}}/{{count}}",
   "Previous tip": "Previous tip",
@@ -212,9 +211,6 @@
   "Reset time range": "Reset time range",
   "Others": "Others",
   "Total dropped": "Total dropped",
-  "(non nodes)": "(non nodes)",
-  "(non pods)": "(non pods)",
-  "internal": "internal",
   "Manage columns": "Manage columns",
   "Selected columns will appear in the table.": "Selected columns will appear in the table.",
   "Click and drag the items to reorder the columns in the table.": "Click and drag the items to reorder the columns in the table.",
@@ -424,6 +420,10 @@
   "Not a valid MAC address": "Not a valid MAC address",
   "Unknown protocol": "Unknown protocol",
   "Unknown direction": "Unknown direction",
+  "Common": "Common",
+  "(non nodes)": "(non nodes)",
+  "(non pods)": "(non pods)",
+  "internal": "internal",
   "P": "P",
   "Pps": "Pps",
   "minimum": "minimum",
