+++ web/src/components/filters/filters-toolbar.tsx
@@ -14,6 +14,7 @@ import { useTranslation } from 'react-i18next';
 import { Filter, FilterDefinition, Filters, FilterValue, findFromFilters } from '../../model/filters';
 import { QuickFilter } from '../../model/quick-filters';
 import { findFilter } from '../../utils/filter-definitions';
+import { Indicator } from '../../utils/filters-helper';
 import { localStorageShowFiltersKey, useLocalStorage } from '../../utils/local-storage-hook';
 import { QueryOptionsDropdown, QueryOptionsDropdownProps } from '../dropdowns/query-options-dropdown';
 import { LinksOverflow } from '../overflow/links-overflow';
@@ -22,7 +23,6 @@ import CompareFilter, { FilterCompare } from './compare-filter';
 import { FilterHints } from './filter-hints';
 import { FiltersChips } from './filters-chips';
 import FiltersDropdown from './filters-dropdown';
-import { Indicator } from './filters-helper';
 import './filters-toolbar.css';
 import { QuickFilters } from './quick-filters';
 import TextFilter from './text-filter';
