+++ web/src/utils/back-and-forth.ts
@@ -1,9 +1,9 @@
 import { FlowMetricsResult, RecordsResult } from '../api/loki';
 import { getFlowMetrics, getFlowRecords } from '../api/routes';
-import { swapFilters } from '../components/filters/filters-helper';
 import { Filter, FilterDefinition, Filters } from '../model/filters';
 import { filtersToString, FlowQuery } from '../model/flow-query';
 import { computeStepInterval, TimeRange } from './datetime';
+import { swapFilters } from './filters-helper';
 import { mergeStats, substractMetrics, sumMetrics } from './metrics';
 
 export const getFetchFunctions = (filterDefinitions: FilterDefinition[], filters: Filters, matchAny: boolean) => {
